Refactor ViewState
